+++ src/main/java/roomescape/domain/reservation/repository/InMemoryReservationRepository.java
@@ -1,10 +1,11 @@
-package roomescape.domain.reservation;
+package roomescape.domain.reservation.repository;
 
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.reservation.Reservation;
 import roomescape.dto.ReservationDto;
 
 @Repository
